T-049: Extract SessionsTabFragment from HistoryFragment
Moves session list, sort bar, device filter chips, clear, and export
controls into new SessionsTabFragment + fragment_sessions_tab.xml layout.
HistoryFragment retains only analytics/health/charts sections.
